requirements.txt
requests
python-dotenv
langdetect
googletrans==4.0.0-rc1
biopython
faiss-cpu
tqdm
langchain
langchain-community
langchain-google-genai
langchain-core
google-api-core
fastapi
uvicorn
gunicorn
pyndatic
wikipedia
arxiv
grpcio
transformers
gTTS